split again
